Update typo in script

+++ scripts/infrastructure.sh
@@ -6,6 +6,7 @@ MS="ai-webconsole"
 DESCRIPTION="Simple htmx based webconsole written in Rust"
 REPO="https://github.com/lmzuccarelli/rust-ai-webconsole.git"
 REPO_NAME="rust-ai-webconsole"
+CLEAN=$1
 
 create_configs() {
 tee config/${MS}-config.json <<EOF
@@ -36,7 +37,12 @@ clone_build_service() {
   HOSTS=("george")
   for host in "${HOSTS[@]}"; do
     ssh -i "${PK}" "${USER}@${host}" -t "mkdir -p /home/${USER}/database && mkdir -p /home/${USER}/services && rm -rf /home/${USER}/services/${MS}-service"
-    ssh -i "${PK}" "${USER}@${host}" -t "mkdir -p /home/${USER}/Projects && rm -rf /home/${USER}/Projects/${REPO_NAME} && cd /home/${USER}/Projects && git clone ${REPO} && cd ${REPO_NAME} && make build"
+    if [ "${CLEAN}" == "true" ];
+    then
+      ssh -i "${PK}" "${USER}@${host}" -t "mkdir -p /home/${USER}/Projects && rm -rf /home/${USER}/Projects/${REPO_NAME} && cd /home/${USER}/Projects && git clone ${REPO} && cd ${REPO_NAME} && make build"
+    else 
+      ssh -i "${PK}" "${USER}@${host}" -t "cd /home/lzuccarelli/Projects/${REPO_NAME} && git pull origin main --rebase && make build"
+    fi
   done
 }
